forced-colors-mode/forced-colors-mode-32.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Forced colors mode - webkit-tap-highlight-color with sys color.</title>
<link rel="help" href="https://www.w3.org/TR/css-color-adjust-1/#forced-colors-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  a {
    -webkit-tap-highlight-color: GrayText;
  }
</style>
<a href="https://www.wikipedia.org" id="link">
  This link color should not be overridden when forced colors mode is enabled.
</a>

<script>
  test(function(){
    assert_not_equals(getComputedStyle(link).webkitTapHighlightColor, "rgba(0, 0, 0, 0.18)");
  }, "Checks hyperlinks are not overridden if a sys color is used.");
</script>
